Add option to pass task names as parameters
Due to the unordered nature of maps in Go, an ordered map implementation had to
be written. This implementation will be move to a separate package. In addition,
since the machines list the tasks available to them, the passed in tasks must be
checked for membership before executing. This will also be moved to a separate
package for easier union/membership checks.